+++ StudentInformationPage/Models/Student.cs
@@ -17,7 +17,6 @@ namespace StudentInformationPage.Models
         [Display(Name = "Last Name")]
         public string? LastName { get; set; }
 
-
         [MaxLength(30, ErrorMessage = "String input must be less than 30")]
         [MinLength(2, ErrorMessage = "String input should be greater than 2")]
         [Display(Name = "Middle Name")]
@@ -26,6 +25,18 @@ namespace StudentInformationPage.Models
         [Display(Name = "Birth Date")]
         [DataType(DataType.Date)]
         public DateTime BirthDate { get; set; }
+
+        [Required]
+        [MaxLength(10, ErrorMessage = "String input must be less than 10")]
+        [MinLength(5, ErrorMessage = "String input must at lease more than 5")]
+        [Display(Name = "Username")]
+        public string? USername { get; set; }
+
+        [Required]
+        [MaxLength(10, ErrorMessage = "String input must be less than 10")]
+        [MinLength(5, ErrorMessage = "String input must at lease more than 5")]
+        [Display(Name = "Password")]
+        public string? Password { get; set; }
 
         [Required]
         [MaxLength(30, ErrorMessage = "String input must be less than 30")]
@@ -47,6 +58,7 @@ namespace StudentInformationPage.Models
         [RegularExpression(@"^(\+\d{1,2}\s?)?\(?\d{3}\)?[\s.-]?\d{3}[\s.-]?\d{4}$", ErrorMessage = "Invalid phone number format.")]
         [DataType(DataType.PhoneNumber)]
         public string? ContactNumber { get; set; }
+
 
     }
 }
