+++ StudentInformationPage/Pages/account-data/Delete.cshtml.cs
@@ -8,17 +8,18 @@ using Microsoft.EntityFrameworkCore;
 using StudentInformationPage.Data;
 using StudentInformationPage.Models;
 
-namespace StudentInformationPage.Pages.student_information
+namespace StudentInformationPage.Pages.account_data
 {
-    public class DetailsModel : PageModel
+    public class DeleteModel : PageModel
     {
         private readonly StudentInformationPage.Data.StudentInformationPageContext _context;
 
-        public DetailsModel(StudentInformationPage.Data.StudentInformationPageContext context)
+        public DeleteModel(StudentInformationPage.Data.StudentInformationPageContext context)
         {
             _context = context;
         }
 
+        [BindProperty]
         public Student Student { get; set; } = default!;
 
         public async Task<IActionResult> OnGetAsync(int? id)
@@ -29,6 +30,7 @@ namespace StudentInformationPage.Pages.student_information
             }
 
             var student = await _context.Student.FirstOrDefaultAsync(m => m.Id == id);
+
             if (student == null)
             {
                 return NotFound();
@@ -38,6 +40,24 @@ namespace StudentInformationPage.Pages.student_information
                 Student = student;
             }
             return Page();
+        }
+
+        public async Task<IActionResult> OnPostAsync(int? id)
+        {
+            if (id == null)
+            {
+                return NotFound();
+            }
+
+            var student = await _context.Student.FindAsync(id);
+            if (student != null)
+            {
+                Student = student;
+                _context.Student.Remove(Student);
+                await _context.SaveChangesAsync();
+            }
+
+            return RedirectToPage("./Index");
         }
     }
 }
