+++ StudentInformationPage/Pages/account-data/Details.cshtml.cs
@@ -8,18 +8,17 @@ using Microsoft.EntityFrameworkCore;
 using StudentInformationPage.Data;
 using StudentInformationPage.Models;
 
-namespace StudentInformationPage.Pages.student_information
+namespace StudentInformationPage.Pages.account_data
 {
-    public class DeleteModel : PageModel
+    public class DetailsModel : PageModel
     {
         private readonly StudentInformationPage.Data.StudentInformationPageContext _context;
 
-        public DeleteModel(StudentInformationPage.Data.StudentInformationPageContext context)
+        public DetailsModel(StudentInformationPage.Data.StudentInformationPageContext context)
         {
             _context = context;
         }
 
-        [BindProperty]
         public Student Student { get; set; } = default!;
 
         public async Task<IActionResult> OnGetAsync(int? id)
@@ -30,7 +29,6 @@ namespace StudentInformationPage.Pages.student_information
             }
 
             var student = await _context.Student.FirstOrDefaultAsync(m => m.Id == id);
-
             if (student == null)
             {
                 return NotFound();
@@ -40,24 +38,6 @@ namespace StudentInformationPage.Pages.student_information
                 Student = student;
             }
             return Page();
-        }
-
-        public async Task<IActionResult> OnPostAsync(int? id)
-        {
-            if (id == null)
-            {
-                return NotFound();
-            }
-
-            var student = await _context.Student.FindAsync(id);
-            if (student != null)
-            {
-                Student = student;
-                _context.Student.Remove(Student);
-                await _context.SaveChangesAsync();
-            }
-
-            return RedirectToPage("./Index");
         }
     }
 }
