+++ StudentInformationPage/Pages/account-data/Index.cshtml
@@ -1,0 +1,102 @@
+﻿@page
+@model StudentInformationPage.Pages.account_data.IndexModel
+
+@{
+    ViewData["Title"] = "Index";
+}
+
+<h1>Index</h1>
+
+<p>
+    <a asp-page="Create">Create New</a>
+</p>
+<table class="table">
+    <thead>
+        <tr>
+            <th>
+                @Html.DisplayNameFor(model => model.Student[0].FirstName)
+            </th>
+            <th>
+                @Html.DisplayNameFor(model => model.Student[0].LastName)
+            </th>
+            <th>
+                @Html.DisplayNameFor(model => model.Student[0].MiddleName)
+            </th>
+            <th>
+                @Html.DisplayNameFor(model => model.Student[0].BirthDate)
+            </th>
+            <th>
+                @Html.DisplayNameFor(model => model.Student[0].USername)
+            </th>
+            <th>
+                @Html.DisplayNameFor(model => model.Student[0].Password)
+            </th>
+            <th>
+                @Html.DisplayNameFor(model => model.Student[0].Course)
+            </th>
+            <th>
+                @Html.DisplayNameFor(model => model.Student[0].YearLevel)
+            </th>
+            <th>
+                @Html.DisplayNameFor(model => model.Student[0].Gender)
+            </th>
+            <th>
+                @Html.DisplayNameFor(model => model.Student[0].Age)
+            </th>
+            <th>
+                @Html.DisplayNameFor(model => model.Student[0].Email)
+            </th>
+            <th>
+                @Html.DisplayNameFor(model => model.Student[0].ContactNumber)
+            </th>
+            <th></th>
+        </tr>
+    </thead>
+    <tbody>
+@foreach (var item in Model.Student) {
+        <tr>
+            <td>
+                @Html.DisplayFor(modelItem => item.FirstName)
+            </td>
+            <td>
+                @Html.DisplayFor(modelItem => item.LastName)
+            </td>
+            <td>
+                @Html.DisplayFor(modelItem => item.MiddleName)
+            </td>
+            <td>
+                @Html.DisplayFor(modelItem => item.BirthDate)
+            </td>
+            <td>
+                @Html.DisplayFor(modelItem => item.USername)
+            </td>
+            <td>
+                @Html.DisplayFor(modelItem => item.Password)
+            </td>
+            <td>
+                @Html.DisplayFor(modelItem => item.Course)
+            </td>
+            <td>
+                @Html.DisplayFor(modelItem => item.YearLevel)
+            </td>
+            <td>
+                @Html.DisplayFor(modelItem => item.Gender)
+            </td>
+            <td>
+                @Html.DisplayFor(modelItem => item.Age)
+            </td>
+            <td>
+                @Html.DisplayFor(modelItem => item.Email)
+            </td>
+            <td>
+                @Html.DisplayFor(modelItem => item.ContactNumber)
+            </td>
+            <td>
+                <a asp-page="./Edit" asp-route-id="@item.Id">Edit</a> |
+                <a asp-page="./Details" asp-route-id="@item.Id">Details</a> |
+                <a asp-page="./Delete" asp-route-id="@item.Id">Delete</a>
+            </td>
+        </tr>
+}
+    </tbody>
+</table>
