--- StudentInformationPage/Pages/student-information/Delete.cshtml
@@ -1,72 +1,0 @@
-﻿@page
-@model StudentInformationPage.Pages.student_information.DeleteModel
-
-@{
-    ViewData["Title"] = "Delete Student";
-}
-
-<div class="d-flex align-items-center justify-content-center vh-100"
-   >
-
-    <div class="glass-card p-4 rounded-4 shadow-lg text-white" style="max-width: 600px; width: 100%;">
-        <h2 class="text-danger fw-bold mb-3">⚠️ Delete Student</h2>
-        <p class="mb-4">Are you sure you want to delete this record? This action <span class="fw-bold text-warning">cannot be undone</span>.</p>
-
-        <div class="glass-inner p-3 rounded-3 mb-4">
-            <dl class="row mb-0">
-                <dt class="col-sm-4 fw-bold">First Name</dt>
-                <dd class="col-sm-8">@Html.DisplayFor(model => model.Student.FirstName)</dd>
-
-                <dt class="col-sm-4 fw-bold">Last Name</dt>
-                <dd class="col-sm-8">@Html.DisplayFor(model => model.Student.LastName)</dd>
-
-                <dt class="col-sm-4 fw-bold" id="email_link">Email</dt>
-                <dd class="col-sm-8">@Html.DisplayFor(model => model.Student.Email)</dd>
-
-                <dt class="col-sm-4 fw-bold">Course</dt>
-                <dd class="col-sm-8">@Html.DisplayFor(model => model.Student.Course)</dd>
-            </dl>
-        </div>
-
-        <form method="post" class="d-flex justify-content-between">
-            <input type="hidden" asp-for="Student.Id" />
-            <button type="submit" class="btn btn-danger px-4 py-2 rounded-3 shadow-sm">
-                Delete
-            </button>
-            <a asp-page="./Index" class="btn btn-outline-light">Back</a>
-
-        </form>
-    </div>
-</div>
-
-<style>
-    .glass-card {
-        background: rgba(255, 255, 255, 0.05);
-        backdrop-filter: blur(12px);
-        border: 1px solid rgba(255, 255, 255, 0.2);
-    }
-
-    .glass-inner {
-        background: rgba(255, 255, 255, 0.08);
-        border: 1px solid rgba(255, 255, 255, 0.15);
-    }
-
-    .glass-input {
-        background: rgba(255, 255, 255, 0.1) !important;
-        border: 1px solid rgba(255, 255, 255, 0.2) !important;
-        border-radius: 8px;
-        color: #EDEDED;
-    }
-
-        .glass-input:focus {
-            outline: none;
-            border-color: #5fffd9c9 !important;
-            box-shadow: 0 0 6px rgba(95, 255, 217, 0.4);
-        }
-
-
-    label {
-        font-weight: 500;
-        font-size: 0.9rem;
-    }
-</style>
