--- StudentInformationPage/Pages/student-information/Details.cshtml
@@ -1,70 +1,0 @@
-﻿@page
-@model StudentInformationPage.Pages.student_information.DetailsModel
-
-@{
-    ViewData["Title"] = "Student Details";
-}
-
-<div class="container py-5">
-    <div class="card glass-card p-4 shadow-lg rounded-4 border-0">
-        <h2 class="text-accent fw-bold mb-4">Student Details</h2>
-
-        <div class="row g-3">
-            <div class="col-12">
-                <label class="detail-label">First Name</label>
-                <p class="detail-value">@Html.DisplayFor(model => model.Student.FirstName)</p>
-            </div>
-
-            <div class="col-12">
-                <label class="detail-label">Last Name</label>
-                <p class="detail-value">@Html.DisplayFor(model => model.Student.LastName)</p>
-            </div>
-
-            <div class="col-12">
-                <label class="detail-label">Middle Name</label>
-                <p class="detail-value">@Html.DisplayFor(model => model.Student.MiddleName)</p>
-            </div>
-
-            <div class="col-12">
-                <label class="detail-label">Birth Date</label>
-                <p class="detail-value">@Html.DisplayFor(model => model.Student.BirthDate)</p>
-            </div>
-
-            <div class="col-12">
-                <label class="detail-label">Course</label>
-                <p class="detail-value">@Html.DisplayFor(model => model.Student.Course)</p>
-            </div>
-
-            <div class="col-12">
-                <label class="detail-label">Year Level</label>
-                <p class="detail-value">@Html.DisplayFor(model => model.Student.YearLevel)</p>
-            </div>
-
-            <div class="col-12">
-                <label class="detail-label">Gender</label>
-                <p class="detail-value">@Html.DisplayFor(model => model.Student.Gender)</p>
-            </div>
-
-            <div class="col-12">
-                <label class="detail-label">Age</label>
-                <p class="detail-value">@Html.DisplayFor(model => model.Student.Age)</p>
-            </div>
-
-            <div class="col-12">
-                <label class="detail-label">Email</label>
-                <p class="detail-value" id="email_link">@Html.DisplayFor(model => model.Student.Email)</p>
-            </div>
-
-            <div class="col-12">
-                <label class="detail-label">Contact Number</label>
-                <p class="detail-value">@Html.DisplayFor(model => model.Student.ContactNumber)</p>
-            </div>
-        </div>
-
-        <div class="mt-4 d-flex justify-content-end gap-3">
-            <a asp-page="./Edit" asp-route-id="@Model.Student.Id" class="btn btn-accent"
-               style="background-color: #ADEBB3; border: none; color: white; border-radius: 8px;">Edit</a>
-            <a asp-page="./Index" class="btn btn-outline-light">Back</a>
-        </div>
-    </div>
-</div>
