--- StudentInformationPage/Pages/student-information/Index.cshtml
@@ -1,81 +1,0 @@
-﻿@page
-@model StudentInformationPage.Pages.student_information.IndexModel
-
-@{
-    ViewData["Title"] = "Index";
-}
-
-
-
-<div class="min-vh-100 d-flex flex-column align-items-center justify-content-start py-5"
-     >
-
-    <div class="card shadow-xl p-4 rounded-4 w-100 glass-card"
-        >
-
-        <!-- Header -->
-        <div class="d-flex justify-content-between align-items-center mb-4">
-            <h2 class="fw-bold m-0 gold-gradient-text">
-                <i class="bi bi-people-fill me-2 gold-gradient-text"></i>
-                Student Information
-            </h2>
-            <a asp-page="Create" class="btn glass-btn px-4 rounded-3 shadow-sm">
-                + Add New
-            </a>
-        </div>
-
-
-        <!-- Table -->
-        <div class="table-responsive">
-            <table class="table glass-table table-hover align-middle text-center mb-0">
-                <thead class="table-light">
-                    <tr>
-                        <th>@Html.DisplayNameFor(model => model.Student[0].FirstName)</th>
-                        <th>@Html.DisplayNameFor(model => model.Student[0].LastName)</th>
-                        <th>@Html.DisplayNameFor(model => model.Student[0].MiddleName)</th>
-                        <th>@Html.DisplayNameFor(model => model.Student[0].BirthDate)</th>
-                        <th>@Html.DisplayNameFor(model => model.Student[0].Course)</th>
-                        <th>@Html.DisplayNameFor(model => model.Student[0].YearLevel)</th>
-                        <th>@Html.DisplayNameFor(model => model.Student[0].Gender)</th>
-                        <th>@Html.DisplayNameFor(model => model.Student[0].Age)</th>
-                        <th>@Html.DisplayNameFor(model => model.Student[0].Email)</th>
-                        <th>@Html.DisplayNameFor(model => model.Student[0].ContactNumber)</th>
-                        <th>Actions</th>
-                    </tr>
-                </thead>
-                <tbody>
-                   
-                    @foreach (var item in Model.Student)
-                    {
-                        <tr>
-                            <td>@Html.DisplayFor(modelItem => item.FirstName)</td>
-                            <td>@Html.DisplayFor(modelItem => item.LastName)</td>
-                            <td>@Html.DisplayFor(modelItem => item.MiddleName)</td>
-                            <td>@Html.DisplayFor(modelItem => item.BirthDate)</td>
-                            <td>@Html.DisplayFor(modelItem => item.Course)</td>
-                            <td>@Html.DisplayFor(modelItem => item.YearLevel)</td>
-                            <td>@Html.DisplayFor(modelItem => item.Gender)</td>
-                            <td>@Html.DisplayFor(modelItem => item.Age)</td>
-                            <td id="email_link">@Html.DisplayFor(modelItem => item.Email)</td>
-                            <td>@Html.DisplayFor(modelItem => item.ContactNumber)</td>
-                            <td>
-                                <a asp-page="./Edit" asp-route-id="@item.Id" class="btn btn-sm glass-btn shadow-sm me-2" style="border-radius: .2rem;" title="Edit">
-                                    <i class="bi bi-pencil-square" style="color: white;"></i>
-                                </a>
-                                <a asp-page="./Details" asp-route-id="@item.Id" class="btn btn-sm glass-btn shadow-sm me-2" style="border-radius: .2rem;" title="Details">
-                                    <i class="bi bi-eye text-info" style="color: white !important;"></i>
-                                </a>
-                                <a asp-page="./Delete" asp-route-id="@item.Id" class="btn btn-sm glass-btn shadow-sm" style="border-radius: .2rem;" title="Delete">
-                                    <i class="bi bi-trash text-danger" style="color: white;"></i>
-                                </a>
-                            </td>
-                        </tr>
-                    }
-                </tbody>
-            </table>
-        </div>
-    </div>
-</div>
-
-<!-- Bootstrap Icons CDN -->
-<link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap-icons@1.11.3/font/bootstrap-icons.css">
